requirements.txt
fastapi
uvicorn[standard]
peewee
pydantic
pandas
openpyxl
python-multipart
python-dotenv
